+++ activestorage/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Allow to detach an attachment when record is not persisted
+
+    *Jacopo Beschi*
+
 *   Use libvips instead of ImageMagick to analyze images when `active_storage.variant_processor = vips`  
 
     *Breno Gazzola*